test(gateway): keep session scope coverage focused
